TB-18/refactor: 패키지 분리
