+++ src/main/java/com/ClubAccount_BE/receipt/adapter/in/web/api/FindReceiptApi.java
@@ -1,9 +1,9 @@
-package com.ClubAccount_BE.receipt.adapter.in.web;
+package com.ClubAccount_BE.receipt.adapter.in.web.api;
 
 import com.ClubAccount_BE.core.meta.LoginUser;
 import com.ClubAccount_BE.core.response.PagingResponse;
-import com.ClubAccount_BE.receipt.adapter.in.web.dto.response.DetailReceiptResponseDto;
-import com.ClubAccount_BE.receipt.adapter.in.web.dto.response.FindReceiptResponseDto;
+import com.ClubAccount_BE.receipt.adapter.in.web.dto.response.ReceiptDetailResponse;
+import com.ClubAccount_BE.receipt.adapter.in.web.dto.response.ReceiptResponse;
 import com.ClubAccount_BE.user.domain.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -16,13 +16,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface FindReceiptApi {
 
     @Operation(summary = "영수증 목록 조회", description = "파싱된 영수증 정보를 조회한다.")
-    PagingResponse<FindReceiptResponseDto> getReceipts(
+    PagingResponse<ReceiptResponse> getReceipts(
             @LoginUser User user,
             @PageableDefault(page = 1, sort = "createdAt", direction = Sort.Direction.ASC) Pageable pageable
     );
 
     @Operation(summary = "영수증 상세 목록 조회", description = "파싱된 영수증 정보를 조회한다.")
-    DetailReceiptResponseDto getReceipt(
+    ReceiptDetailResponse getReceipt(
             @LoginUser User user,
             @PathVariable("receiptId") Long receiptId
     );
